ATTENTION, this commit breaks the Trigger code API.
With this commit it is possible to run triggers with button click.
I am planing to use this to load worlds.
This commit moves the Trigger code de/serialize to Trigger
interface, and marks proper methods static. Since it is not
possible to acces instance fields in those methods, some changes
were required in API itself.