WIP Add tag based rendering
It crashes with object creation, and only tested on static objects.

I realized the camera tags should not be explicit but inferred.
Render steps should have tags and camera, and they together should define
culling, for the time being, I am pretending that I infer the tags,
but I actually set it in the interface.