Move Texture load out of material load
This is a step to make model asset loading parallel.
Since the flow goes as
model -> mesh ->material -> texture -> GPU
and GPU can't be parallel in OpenGL, we have to move the
texture part out up to model.

Same goes for models and meshes, but they already have partial support
because of the cereal implementation.