WIP material based rendering order
TODO: breaks animations

This is supposed to improve performance while allowing
materials for copy objects to be altered by its own.

The fix part works, but performance improve doesn't,
and thats because old order was random, and this ordering
is also partially random. We need to order it in some way
to actually see relevant changes improving the performance.